requirements.txt
# Usage: pip install -r requirements.txt
pyyaml==6.0.1
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.2.0
scipy==1.13.0
tqdm==4.65.2
pandas==2.2.2
scikit-image==0.23.2
matplotlib==3.8.4
tensorboard==2.16.2
plotly==5.21.0
timm